+++ test/command-help.js
@@ -3,14 +3,14 @@ var stdout = require('test-console').stdout;
 var cli = require('../lib');
 
 describe('Command help', () => {
+    var inspect;
+    beforeEach(() => inspect = stdout.inspect());
+    afterEach(() => inspect.restore());
+
     it('should show help', () => {
-        var output;
+        cli.command('test', false).run(['--help']);
 
-        cli.command('test', false, {
-            infoOptionAction: res => output = res
-        }).run(['--help']);
-
-        assert.equal(output, [
+        assert.equal(inspect.output, [
             'Usage:',
             '',
             '    \u001b[36mtest\u001b[39m [\u001b[33moptions\u001b[39m]',
@@ -18,17 +18,14 @@ describe('Command help', () => {
             'Options:',
             '',
             '    \u001b[33m-h\u001b[39m, \u001b[33m--help\u001b[39m                 Output usage information',
+            '',
             ''
         ].join('\n'));
     });
 
     it('should show help all cases', () => {
-        var output;
-
         cli
-            .command('test', '[qux]', {
-                infoOptionAction: res => output = res
-            })
+            .command('test', '[qux]')
             .description('Test description')
             .option('-f, --foo', 'Foo')
             .option('--bar <baz>', 'Bar')
@@ -36,7 +33,7 @@ describe('Command help', () => {
             .end()
             .run(['--help']);
 
-        assert.equal(output, [
+        assert.equal(inspect.output, [
             'Test description',
             '',
             'Usage:',
@@ -52,24 +49,21 @@ describe('Command help', () => {
             '        \u001b[33m--bar\u001b[39m <baz>            Bar',
             '    \u001b[33m-f\u001b[39m, \u001b[33m--foo\u001b[39m                  Foo',
             '    \u001b[33m-h\u001b[39m, \u001b[33m--help\u001b[39m                 Output usage information',
+            '',
             ''
         ].join('\n'));
     });
 
     it('should show help for nested command', () => {
-        var output;
-
         cli
             .command('test', '[qux]')
             .option('-f, --foo', 'Foo')
-            .command('nested', '[nested-arg]', {
-                infoOptionAction: res => output = res
-            })
+            .command('nested', '[nested-arg]')
             .option('--bar <baz>', 'Bar')
             .end()
             .run(['nested', '--help']);
 
-        assert.equal(output, [
+        assert.equal(inspect.output, [
             'Usage:',
             '',
             '    \u001b[36mtest nested\u001b[39m \u001b[35m[nested-arg]\u001b[39m [\u001b[33moptions\u001b[39m]',
@@ -78,6 +72,7 @@ describe('Command help', () => {
             '',
             '        \u001b[33m--bar\u001b[39m <baz>            Bar',
             '    \u001b[33m-h\u001b[39m, \u001b[33m--help\u001b[39m                 Output usage information',
+            '',
             ''
         ].join('\n'));
     });
@@ -91,16 +86,13 @@ describe('Command help', () => {
     });
 
     it('should show help message when Command#showHelp called', function() {
-        var inspect = stdout.inspect();
-
         var command = cli
             .command('test', '[qux]')
             .option('-f, --foo', 'Foo');
 
         command.showHelp();
-        inspect.restore();
 
-        assert.equal(inspect.output.join(''), [
+        assert.equal(inspect.output, [
             'Usage:',
             '',
             '    \u001b[36mtest\u001b[39m \u001b[35m[qux]\u001b[39m [\u001b[33moptions\u001b[39m]',